[Oztechan/CCC#1764] Move common used classes into core modules
