+++ common/core/database/src/commonMain/kotlin/com/oztechan/ccc/common/core/database/error/DatabaseException.kt
@@ -1,0 +1,3 @@
+package com.oztechan.ccc.common.core.database.error
+
+class DatabaseException(cause: Throwable) : Throwable(cause)
