+++ common/core/network/src/commonMain/kotlin/com/oztechan/ccc/common/core/network/error/EmptyParameterException.kt
@@ -1,0 +1,3 @@
+package com.oztechan.ccc.common.core.network.error
+
+class EmptyParameterException : Exception("parameter can not be empty")
