+++ common/core/network/src/commonMain/kotlin/com/oztechan/ccc/common/core/network/error/NetworkException.kt
@@ -1,0 +1,3 @@
+package com.oztechan.ccc.common.core.network.error
+
+class NetworkException(cause: Throwable) : Throwable(cause)
