+++ common/core/network/src/commonMain/kotlin/com/oztechan/ccc/common/core/network/error/TimeoutException.kt
@@ -1,0 +1,3 @@
+package com.oztechan.ccc.common.core.network.error
+
+class TimeoutException(cause: Throwable) : Throwable(cause)
