+++ common/core/network/src/commonMain/kotlin/com/oztechan/ccc/common/core/network/error/UnknownNetworkException.kt
@@ -1,0 +1,3 @@
+package com.oztechan.ccc.common.core.network.error
+
+class UnknownNetworkException(cause: Throwable) : Throwable(cause)
