--- common/src/commonMain/kotlin/com/oztechan/ccc/common/error/DatabaseException.kt
@@ -1,3 +1,0 @@
-package com.oztechan.ccc.common.error
-
-internal class DatabaseException(cause: Throwable) : Throwable(cause)
