--- common/src/commonMain/kotlin/com/oztechan/ccc/common/error/EmptyParameterException.kt
@@ -1,3 +1,0 @@
-package com.oztechan.ccc.common.error
-
-internal class EmptyParameterException : Exception("parameter can not be empty")
