--- common/src/commonMain/kotlin/com/oztechan/ccc/common/error/ModelMappingException.kt
@@ -1,3 +1,0 @@
-package com.oztechan.ccc.common.error
-
-internal class ModelMappingException(cause: Throwable) : Throwable(cause)
