--- common/src/commonMain/kotlin/com/oztechan/ccc/common/error/NetworkException.kt
@@ -1,3 +1,0 @@
-package com.oztechan.ccc.common.error
-
-internal class NetworkException(cause: Throwable) : Throwable(cause)
