--- common/src/commonMain/kotlin/com/oztechan/ccc/common/error/TimeoutException.kt
@@ -1,3 +1,0 @@
-package com.oztechan.ccc.common.error
-
-internal class TimeoutException(cause: Throwable) : Throwable(cause)
