--- common/src/commonMain/kotlin/com/oztechan/ccc/common/error/UnknownNetworkException.kt
@@ -1,3 +1,0 @@
-package com.oztechan.ccc.common.error
-
-internal class UnknownNetworkException(cause: Throwable) : Throwable(cause)
