+++ common/src/commonMain/kotlin/com/oztechan/ccc/common/service/backend/BackendApiServiceImpl.kt
@@ -3,7 +3,7 @@ package com.oztechan.ccc.common.service.backend
 import co.touchlab.kermit.Logger
 import com.oztechan.ccc.common.core.network.api.backend.BackendApi
 import com.oztechan.ccc.common.core.network.mapper.toExchangeRateModel
-import com.oztechan.ccc.common.service.BaseNetworkService
+import com.oztechan.ccc.common.core.network.base.BaseNetworkService
 import kotlinx.coroutines.CoroutineDispatcher
 
 internal class BackendApiServiceImpl(
