+++ common/src/commonMain/kotlin/com/oztechan/ccc/common/service/free/FreeApiServiceImpl.kt
@@ -2,8 +2,8 @@ package com.oztechan.ccc.common.service.free
 
 import co.touchlab.kermit.Logger
 import com.oztechan.ccc.common.core.network.api.free.FreeApi
+import com.oztechan.ccc.common.core.network.base.BaseNetworkService
 import com.oztechan.ccc.common.core.network.mapper.toExchangeRateModel
-import com.oztechan.ccc.common.service.BaseNetworkService
 import kotlinx.coroutines.CoroutineDispatcher
 
 internal class FreeApiServiceImpl(
