+++ common/src/commonMain/kotlin/com/oztechan/ccc/common/service/premium/PremiumApiServiceImpl.kt
@@ -5,8 +5,8 @@ package com.oztechan.ccc.common.service.premium
 
 import co.touchlab.kermit.Logger
 import com.oztechan.ccc.common.core.network.api.premium.PremiumApi
+import com.oztechan.ccc.common.core.network.base.BaseNetworkService
 import com.oztechan.ccc.common.core.network.mapper.toExchangeRateModel
-import com.oztechan.ccc.common.service.BaseNetworkService
 import kotlinx.coroutines.CoroutineDispatcher
 
 internal class PremiumApiServiceImpl(
