+++ common/src/commonTest/kotlin/com/oztechan/ccc/common/service/FreeApiServiceTest.kt
@@ -1,10 +1,10 @@
 package com.oztechan.ccc.common.service
 
 import com.oztechan.ccc.common.core.network.api.free.FreeApi
+import com.oztechan.ccc.common.core.network.error.UnknownNetworkException
 import com.oztechan.ccc.common.core.network.mapper.toExchangeRateModel
 import com.oztechan.ccc.common.core.network.model.Conversion
 import com.oztechan.ccc.common.core.network.model.ExchangeRate
-import com.oztechan.ccc.common.error.UnknownNetworkException
 import com.oztechan.ccc.common.service.free.FreeApiService
 import com.oztechan.ccc.common.service.free.FreeApiServiceImpl
 import com.oztechan.ccc.test.BaseSubjectTest
